修改 add 函数返回值

+++ test.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int add(int a, int b)
+long add(int a, int b)
 {
 	return a + b;
 }
